requirements.txt
librosa
numpy
scipy
matplotlib
pandas
python-dotenv
ossapi
sklearn
aiohttp